refactor: Update BACKEND_PROVIDER in main.ts to use prisma
This commit updates the `BACKEND_PROVIDER` variable in the `main.ts` file to use the value `prisma` instead of `supabase`. This change aligns the backend provider configuration with the recent updates made to the user-related components and middleware, which now use the new `userStore`. Additionally, it reflects the shift in the project's direction towards utilizing Prisma as the backend provider.

#prisma #config